@@ -3,5 +3,5 @@
 </script>
 
 <div class="w-72 h-52 flex justify-center items-center bg-teal-50 rounded-3xl">
-	<p class="font-inter font-semibold text-3xl text-teal-800">{card}</p>
+	<p class="font-semibold text-3xl text-teal-800">{card}</p>
 </div>
