@@ -12,6 +12,6 @@
 
 <div class="inline-block">
 	<div class="w-16 h-16 bg-teal-500 flex justify-center items-center rounded-full inline-block">
-        <p class="text-white font-inter font-medium text-3xl">{value}</p>
+        <p class="text-white text-3xl">{value}</p>
 	</div>
 </div>
