@@ -3,12 +3,7 @@
 </script>
 
 <svelte:head>
-	<link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin="true" />
-	<link
-		href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap"
-		rel="stylesheet"
-	/>
+    <title>Casquette game</title>
 </svelte:head>
 
 <slot />
