@@ -17,8 +17,8 @@
 </script>
 
 <div class="flex flex-col justify-center">
-	<p class="mt-14 text-center font-inter font-medium text-2xl">{username}</p>
-	<p class="mt-3 text-center font-inter font-medium text-4xl">Scores</p>
+	<p class="mt-14 text-center text-2xl">{username}</p>
+	<p class="mt-3 text-center text-4xl">Scores</p>
 
 	{#each score as elem}
 		<TeamScores team={elem.team} scores={elem.points} />
@@ -26,7 +26,7 @@
 
 	{#if isHost && round < 3}
 		<div class="flex justify-center absolute bottom-28 w-full">
-			<Button on:click={handleClick} className="font-semibold">Next</Button>
+			<Button on:click={handleClick}>Next</Button>
 		</div>
 	{/if}
 </div>
