@@ -1,17 +1,17 @@
 import adapter from '@sveltejs/adapter-static';
-import preprocess from 'svelte-preprocess';
+import { vitePreprocess } from '@sveltejs/kit/vite';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	kit: {
-		adapter: adapter({
-			fallback: 'app.html',
-		}),
-		alias: {
-			'@': 'src',
-		},
-	},
-	preprocess: [preprocess({ postcss: true })],
+    preprocess: vitePreprocess(),
+    kit: {
+        adapter: adapter({
+            fallback: 'app.html',
+        }),
+        alias: {
+            '@': 'src',
+        }
+    }
 };
 
 export default config;
